// apps/mobile/components/task-edit/TaskEditFieldRenderer.types.ts
export type PickerOption<T extends string> = {
    value: T | '';
    label: string;
};